Lazy loading internal modules and some dependencies for faster load times. This essentially defers loading child modules until they are explicitly required (using the technique described [here](http://www.typescriptlang.org/docs/handbook/modules.html#optional-module-loading-and-other-advanced-loading-scenarios)). For example none of the `auth/*` sources will be loaded until `admin.auth()` is invoked. This also ensures that only the child modules used by the developers get loaded into the Node.js process.

## Experimental results

Test code:
```
const start = Date.now();
require('firebase-admin');
const end = Date.now();
console.log('Load time:', (end-start), 'ms');
```

### MacBook Pro (Core i7 2.2 GHz, 16GB RAM)

Before (ms): 154, 151, 151, 153, 153 (mean = 152.4ms)
After (ms): 18, 18, 17, 18, 18 (mean = 17.8ms)
Speed up of initial load: 88.32%

### Ubuntu 16.04 (Core i7 2.8GHz, 16GB RAM)

Before (ms): 61, 59, 59, 59, 59 (mean = 59.4 ms)
After (ms): 4, 5, 5, 4, 5 (mean = 4.6 ms)
Speed up of initial load: 92.25%