@hiranya911 2018-05-14
Done. Removed package-lock changes from the PR.

Also FYI @laurenzlong. I think Functions is likely to be the biggest beneficiary of this change.